requirements.txt
pydantic>=2.0.0
pulp>=2.7.0
pyyaml>=6.0
openai>=1.0.0
numpy>=1.24.0
